gnn_sql_project/requirements.txt
networkx
pandas
psycopg2-binary
tabulate
torch
torch-geometric